requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-storage
pandas
openpyxl
apache-airflow-providers-google
google-cloud-bigquery